Fix Testcontainers tests & add comments
