+++ src/main/resources/application-production.properties
@@ -1,6 +1,6 @@
 # this file contains settings for production environments
 
-# database config
+# prod database config
 spring.datasource.username=${POSTGRES_USER}
 spring.datasource.password=${POSTGRES_PASSWORD}
 spring.datasource.url=${POSTGRES_DB}
