+++ src/main/resources/application.properties
@@ -2,6 +2,12 @@
 spring.config.import=optional:file:.env[.properties]
 spring.application.name=marketplace-backend
 
+# dev database config
+spring.datasource.username=${POSTGRES_USER}
+spring.datasource.password=${POSTGRES_PASSWORD}
+spring.datasource.url=${POSTGRES_DB}
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
+
 # disable unwanted automatic database schema updates by hibernate
 spring.jpa.hibernate.ddl-auto=none
